fastapi_postgres_app/pyproject.toml
[tool.poetry]
name = "dirty-c"
version = "0.1.0"
description = "devops_adventure"
authors = ["Diogo Vieira <diogo.romero.paz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
